[MCJ-1367] refactor: auth/user DTO 패키지를 단일 dto 구조로 통일
