+++ src/main/kotlin/com/moongchijang/application/auth/dto/AccessTokenReissueResult.kt
@@ -1,10 +1,7 @@
 package com.moongchijang.application.auth.dto
 
-import com.moongchijang.application.auth.dto.response.AccessTokenResponse
-
 data class AccessTokenReissueResult(
 
     val response: AccessTokenResponse,
     val refreshToken: String,
 )
-
