+++ src/main/kotlin/com/moongchijang/application/auth/dto/AuthLoginResult.kt
@@ -1,7 +1,5 @@
 package com.moongchijang.application.auth.dto
 
-import com.moongchijang.application.auth.dto.response.AuthLoginResponse
-
 data class AuthLoginResult(
 
     val response: AuthLoginResponse,
