+++ src/main/kotlin/com/moongchijang/application/user/dto/NicknameAvailabilityResponse.kt
@@ -1,4 +1,4 @@
-package com.moongchijang.application.user.dto.response
+package com.moongchijang.application.user.dto
 
 import io.swagger.v3.oas.annotations.media.Schema
